+++ Lab_3_2/src/lab32/model/IWeight.java
@@ -2,6 +2,6 @@ package lab32.model;
 
 public interface IWeight {
 
-    float weight();
+    double weight();
 
 }
